Schedules are now read from a directory all at once
Schedules are now read from a directory, and all available schedules are
read at once. Right now this is precarious because the files are not
marked or removed after being read once, so on running the program
multiple times events will be duplicated onto a the calendar.
This will be changed soon to either delete the file or otherwise mark it
as read.

There is also now the option for a debug flag when running the program.
It really just gives some control flow when trying out different things that
get posted to the TEST calendar.

Position is now correctly parsed, so now the events are correctly labeled as
what the shift is. It's kind of janky but not sure how else to change it.

TODO:
   - Make it so that schedules are deleted from directory after being parsed
   - Create a method to clear events off of the TEST calendar
   - Add getopt or argparse, the current if statement is gross but got put in
	end of day. I'm okay with it for right now